rename: service 패키지를 application 패키지로 수정
