+++ src/main/java/roomescape/application/dto/request/member/MemberInfo.java
@@ -1,0 +1,4 @@
+package roomescape.application.dto.request.member;
+
+public record MemberInfo(Long id) {
+}
