+++ src/main/java/roomescape/application/dto/response/member/MemberResponse.java
@@ -1,0 +1,9 @@
+package roomescape.application.dto.response.member;
+
+import roomescape.domain.member.Member;
+
+public record MemberResponse(Long id, String name) {
+    public static MemberResponse from(Member member) {
+        return new MemberResponse(member.getId(), member.getName());
+    }
+}
