+++ src/main/java/roomescape/application/dto/response/theme/ThemeResponse.java
@@ -1,0 +1,15 @@
+package roomescape.application.dto.response.theme;
+
+import roomescape.domain.reservationdetail.Theme;
+
+public record ThemeResponse(Long id, String name, String description, String thumbnail) {
+
+    public static ThemeResponse from(Theme theme) {
+        return new ThemeResponse(
+                theme.getId(),
+                theme.getName(),
+                theme.getDescription(),
+                theme.getThumbnail()
+        );
+    }
+}
