+++ src/main/java/roomescape/application/policy/RankingPolicy.java
@@ -1,0 +1,11 @@
+package roomescape.application.policy;
+
+import java.time.LocalDate;
+
+public interface RankingPolicy {
+    LocalDate getStartDateAsString();
+
+    LocalDate getEndDateAsString();
+
+    int exposureSize();
+}
