+++ src/main/java/roomescape/application/policy/WeeklyRankingPolicy.java
@@ -1,0 +1,23 @@
+package roomescape.application.policy;
+
+import java.time.LocalDate;
+
+public class WeeklyRankingPolicy implements RankingPolicy {
+
+    @Override
+    public LocalDate getStartDateAsString() {
+        return LocalDate.now()
+                .minusDays(8);
+    }
+
+    @Override
+    public LocalDate getEndDateAsString() {
+        return LocalDate.now()
+                .minusDays(1);
+    }
+
+    @Override
+    public int exposureSize() {
+        return 10;
+    }
+}
