--- src/main/java/roomescape/service/dto/request/member/MemberInfo.java
@@ -1,4 +1,0 @@
-package roomescape.service.dto.request.member;
-
-public record MemberInfo(Long id) {
-}
