--- src/main/java/roomescape/service/dto/response/member/MemberResponse.java
@@ -1,9 +1,0 @@
-package roomescape.service.dto.response.member;
-
-import roomescape.domain.member.Member;
-
-public record MemberResponse(Long id, String name) {
-    public static MemberResponse from(Member member) {
-        return new MemberResponse(member.getId(), member.getName());
-    }
-}
