--- src/main/java/roomescape/service/dto/response/theme/ThemeResponse.java
@@ -1,15 +1,0 @@
-package roomescape.service.dto.response.theme;
-
-import roomescape.domain.reservationdetail.Theme;
-
-public record ThemeResponse(Long id, String name, String description, String thumbnail) {
-
-    public static ThemeResponse from(Theme theme) {
-        return new ThemeResponse(
-                theme.getId(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnail()
-        );
-    }
-}
