--- src/main/java/roomescape/service/dto/response/time/ReservationTimeResponse.java
@@ -1,22 +1,0 @@
-package roomescape.service.dto.response.time;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonFormat.Shape;
-import java.time.LocalTime;
-import roomescape.domain.reservationdetail.ReservationTime;
-
-public record ReservationTimeResponse(Long id, LocalTime startAt) {
-
-    public static ReservationTimeResponse from(ReservationTime time) {
-        return new ReservationTimeResponse(
-                time.getId(),
-                time.getStartAt()
-        );
-    }
-
-    @Override
-    @JsonFormat(shape = Shape.STRING, pattern = "HH:mm")
-    public LocalTime startAt() {
-        return startAt;
-    }
-}
