--- src/main/java/roomescape/service/policy/RankingPolicy.java
@@ -1,11 +1,0 @@
-package roomescape.service.policy;
-
-import java.time.LocalDate;
-
-public interface RankingPolicy {
-    LocalDate getStartDateAsString();
-
-    LocalDate getEndDateAsString();
-
-    int exposureSize();
-}
