--- src/main/java/roomescape/service/policy/WeeklyRankingPolicy.java
@@ -1,23 +1,0 @@
-package roomescape.service.policy;
-
-import java.time.LocalDate;
-
-public class WeeklyRankingPolicy implements RankingPolicy {
-
-    @Override
-    public LocalDate getStartDateAsString() {
-        return LocalDate.now()
-                .minusDays(8);
-    }
-
-    @Override
-    public LocalDate getEndDateAsString() {
-        return LocalDate.now()
-                .minusDays(1);
-    }
-
-    @Override
-    public int exposureSize() {
-        return 10;
-    }
-}
