+++ src/main/java/roomescape/web/api/AdminReservationController.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.reservation.Reservation;
-import roomescape.service.ReservationService;
-import roomescape.service.dto.request.reservation.ReservationRequest;
-import roomescape.service.dto.request.reservation.ReservationSearchCondition;
-import roomescape.service.dto.response.reservation.ReservationResponse;
+import roomescape.application.ReservationService;
+import roomescape.application.dto.request.reservation.ReservationRequest;
+import roomescape.application.dto.request.reservation.ReservationSearchCondition;
+import roomescape.application.dto.response.reservation.ReservationResponse;
 
 @RestController
 @RequiredArgsConstructor
