+++ src/main/java/roomescape/web/api/MemberThemeController.java
@@ -5,9 +5,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.service.policy.WeeklyRankingPolicy;
-import roomescape.service.ThemeService;
-import roomescape.service.dto.response.theme.ThemeResponse;
+import roomescape.application.policy.WeeklyRankingPolicy;
+import roomescape.application.ThemeService;
+import roomescape.application.dto.response.theme.ThemeResponse;
 
 @RestController
 @RequiredArgsConstructor
