+++ src/main/java/roomescape/web/config/WebMvcConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.service.security.JwtProvider;
+import roomescape.application.security.JwtProvider;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
